review(!265): read DIS type byte as unsigned; document mtu guard invariant
- S1: parse the DIS AD type byte as unsigned so crafted values above
  0x7F no longer silently skip local name / appearance handling
  (aligns with the unsigned reads in the same file)
- S2: document that mtu is clamped to [SSAP_STACK_MTU_DEFAULT,
  SSAP_STACK_MTU_MAX] at both write sites and the V10 lower-bound check
  is defense-in-depth against that invariant being relaxed

+++ utils/src/sle_service_data.cpp
@@ -1147,7 +1147,7 @@ void SlePeripheralDevice::ParseSleDisData()
     int type = -1;
     size_t length = 0;
     while ((i + SLE_ADV_DEVICE_INFO_MIN_LEN) < disServiceData.size()) {
-        type = disServiceData[i++];
+        type = static_cast<uint8_t>(disServiceData[i++]);
         length = static_cast<uint8_t>(disServiceData[i++]);
         if (length <= 0 || i + length > disServiceData.size()) {
             break;
